refactor: split olf.cli into per-group command modules
Move each Typer command group out of the single 731-line cli.py into
olf/commands/<group>.py: contracts, catalog, floe, artifacts, layers,
revision, superset, openmetadata, k8s, e2e, smoke, release. cli.py now
only constructs the root app, registers each group, and defines
version/main() — same 12 group names, same 22 commands, same flags,
defaults, and exit codes.

_repo_root()/_truthy() moved to olf.config (repo_root()/truthy()) as
generic env helpers reused across command modules. deploy_superset_reports
and deploy_openmetadata_metadata now live next to their own Typer command
in commands/superset.py and commands/openmetadata.py; commands/artifacts.py
imports them for deploy-optional-layers.

Update test_cli.py's monkeypatch targets to the new module paths
(olf.commands.artifacts.*, olf.commands.catalog.*) — patching a bare-name
import only works at the module where the call happens, not at the old
cli.py location. No runner.invoke(...) assertion changed.
