+++ tools/olf/olf/commands/artifacts.py
@@ -1,0 +1,96 @@
+"""Object-storage artifact helpers."""
+
+from __future__ import annotations
+
+import os
+from pathlib import Path
+
+import typer
+
+from olf import config
+from olf import layers as layers_module
+from olf.commands._shared import fail
+from olf.commands.openmetadata import deploy_openmetadata_metadata
+from olf.commands.superset import deploy_superset_reports
+
+app = typer.Typer(help="Object-storage artifact helpers.")
+
+
+@app.command("deploy-optional-layers")
+def artifacts_deploy_optional_layers() -> None:
+    """Deploy artifacts for enabled optional platform layers."""
+    layers_module.deploy_enabled_artifacts(
+        os.environ,
+        deploy_reports=deploy_superset_reports,
+        deploy_metadata=deploy_openmetadata_metadata,
+        report=lambda message: typer.echo(f"==> {message}"),
+    )
+
+
+@app.command("upload-manifests")
+def artifacts_upload_manifests(
+    via: str = typer.Option(
+        "port-forward",
+        "--via",
+        help="'port-forward' for in-cluster S3-compatible storage, 'direct' for cloud S3.",
+    ),
+    manifest_root: str = typer.Option(
+        "",
+        "--manifest-root",
+        help="Rendered-manifest directory for --via direct (default: .tmp/floe-runtime/aws/manifests).",
+    ),
+    runtime_root: str = typer.Option(
+        "",
+        "--runtime-root",
+        help="Rendered Floe runtime artifact root containing configs/, profiles/, and manifests/.",
+    ),
+) -> None:
+    """Publish product Floe runtime artifacts to the operational artifact bucket."""
+    from olf import s3
+
+    repo_root = config.repo_root()
+    namespace = config.namespace()
+    bucket = config.env("OPENLAKEFORGE_OPS_BUCKET_NAME") or config.env("OPENLAKEFORGE_ARTIFACT_BUCKET_NAME")
+    if not bucket:
+        raise typer.Exit(code=fail("no ops/artifact bucket resolved from the contract environment."))
+
+    if via == "direct":
+        if runtime_root:
+            uploads = s3.discover_runtime_artifacts(Path(runtime_root))
+        else:
+            root = Path(manifest_root) if manifest_root else repo_root / ".tmp/floe-runtime/aws/manifests"
+            uploads = s3.discover_runtime_manifests(root)
+        if not uploads:
+            root = runtime_root or manifest_root or str(repo_root / ".tmp/floe-runtime/aws/manifests")
+            raise typer.Exit(code=fail(f"no rendered Floe artifacts found under {root}."))
+        s3.upload_direct(bucket, uploads, region=config.env("OPENLAKEFORGE_STORAGE_REGION"))
+    elif via == "port-forward":
+        if runtime_root:
+            uploads = s3.discover_runtime_artifacts(Path(runtime_root))
+        else:
+            uploads = s3.discover_tracked_manifests(repo_root)
+        if not uploads:
+            raise typer.Exit(code=fail("no generated product Floe artifacts found. Run 'make floe-manifest' first."))
+        secret_name = config.env("OPENLAKEFORGE_STORAGE_CREDENTIALS_SECRET_NAME")
+        service = config.env("OPENLAKEFORGE_STORAGE_S3_SERVICE_NAME", "seaweedfs-s3")
+        remote_port = int(config.env("OPENLAKEFORGE_STORAGE_S3_SERVICE_PORT", "8333"))
+        from olf import k8s
+
+        s3.upload_via_port_forward(
+            bucket,
+            uploads,
+            service=service,
+            remote_port=remote_port,
+            namespace=namespace,
+            access_key_id=k8s.secret_value(
+                secret_name, config.env("OPENLAKEFORGE_STORAGE_ACCESS_KEY_ID_KEY", "AWS_ACCESS_KEY_ID"), namespace
+            ),
+            secret_access_key=k8s.secret_value(
+                secret_name,
+                config.env("OPENLAKEFORGE_STORAGE_SECRET_ACCESS_KEY_KEY", "AWS_SECRET_ACCESS_KEY"),
+                namespace,
+            ),
+            region=config.env("OPENLAKEFORGE_STORAGE_REGION", "us-east-1"),
+        )
+    else:
+        raise typer.Exit(code=fail(f"unknown --via mode: {via!r} (expected 'port-forward' or 'direct')."))
